Operator: Initial implementation
We add an Operator to deploy the components via the OpenPERouter resource.
This is mostly a copy from the metallb-operator, where we consume the helm chart
directly with slight modifications, and adjust for this project + having it in the same repo.
In addition, we add a CI lane that deploys the components via the Operator
and runs the usual e2etests, to verify everything is deployed correctly.

Signed-off-by: Ori Braunshtein <obraunsh@redhat.com>