Enforce EBGP
EBGP has different and more predictable dynamics compared to ibgp. Given
that the router asn is arbitrary, there is no point today in allowing
ibgp.

Note that we are not validating if a user configures the host ASN with
the same ASN of the top of the rack, as it falls outside of the router
scope.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>