Rename the interfaces configuration function
In order to make future the refactoring that will follow simples, this
patch renames the `configureInterfaces` function to `configureHost` to
reflect better that this is overall configuration of the host network.

With this, it will be easier to differentiate between types related to
this overall configuration, and types related specifically to NIC
configuration.

Signed-off-by: Petr Horacek <phoracek@redhat.com>