clab, kind: nodes with extra L2 network interface
This commit refactors the 00-environment.sh script to be a general
purpose script to create an L2 network to interconnect all the cluster's
nodes.

When ran on multi-cluster scenario, we request an additional network to
be plumbed into both cluster's KinD nodes - this network will be used in
follow-up commits as a dedicated network to be used for KubeVirt live
migration.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>