demo, multi-cluster: install whereabouts ipam
We need to setup a dedicated migration network for both clusters to use.
It requires IPAM across the network - in both clusters.

Whereabouts CNI is one of the alternatives recommended by the dedicated
migration network KubeVirt feature, hence, we're going along with it.

This commit installs whereabouts in both KinD clusters.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>